Allow empty values (#51)
Allows empty values, by representing them as tombstones.

This introduces two new types, KeyValueDeletable and ValueDeletable. 
ValueDeletable is a value that may be tombstoned. Although we could 
represent this with Option<Bytes>, we also use that for values that don't 
exist (e.g. resulting from a get() for a missing key). I decided that having 
a dedicated type for tombstone-able values is more readable and makes the 
compiler ensure that we don't mix the two meanings up. KeyValueDeletable is 
a key-value pair that uses ValueDeletable instead of a raw Bytes.

At the block level, tombstones are represented by setting the length prefix 
of the value to 0xFFFFFFFF. This means that the maximum length of a value 
will be one byte shorter than it would have been before.