Add manifest design doc (#43)
Define a design for a manifest. The manifest is responsible for persisting `DbState` in object storage so a process may recover its state on restart. The manifest also contains snapshots so clients may define a single point in time for reads to run off of.

This is the result of discussions from #24 and #39. It features a manifest design that:

- Has a sequential, contiguous write ahead log (WAL)
- Supports parallel writes
- Supports read snapshots
- Does not lock the manifest on every WAL write
- Should work with transactions
- Should work with MVCC and branching