ensure we write to memtable while holding state lock
This patch changes the write path to ensure that we write to the memtable
while holding the read lock on dbstate. This protects against the following
race condition:
1. writer selects the current memtable for writing
2. flush thread freezes the current memtable
3. flush thread flushes the current memtable
4. writer writes to the flushed memtable

To acheive this, this change also refactors the memtable api a bit. put
and put_optimistic are combined into just put, which does not block on
flush. The writer blocks on flush using a call to a new method called
await_flush, which waits for the flush notification.