Track the size of the data in memtable
This commit adds support to the memtable for tracking the current
memtable size in bytes. This actually required a bit of a refactor
of dbstate:

Wraps MemTable in WriteableMemTable. All puts/deletes must go through
WriteableMemtable, which also tracks size from put/delete. put/delete
now require a mutable reference (SkipMap does not require a mutable
ref to insert, (I think because its impl uses unsafe)).

Because WriteableMemtable requires a mutable reference to put/del, it
can no longer be accessed through an `Arc`. It also doesn't make sense
for it to derive `Clone`. Therefore this change refactors the dbstate
structure and mutation model as follows:
- `DbState` is split into `DbState` and `CompactedDbState`. `DbState`
   contains the current mutable memtable and an Arc pointing to
   `CompactedDbState` which contains immutable memtables and l0.
- When the memtable is frozen, it is pulled out of `WriteableMemTable`
  and added to the immutable list, and the current memtable is set
  to a new `WriteableMemtable`. The mutation to the current memtable
  is made directly to dbstate (no snapshot) while holding the
  write-lock
- All mutations to compacted dbstate create a clone, mutate it, and
  reset the Arc.