fix wal to memtabe recovery and move writes to flush thread
fixes a bug in recovery of the imm memtables from the wal on db
startup. In particular, we were previously taking a snapshot before
replaying the wal, and then we were updating the db state with the
snapshot, overriding the changes made from the replay. With this
fix we update the dbstate in place as we go along.

this patch also moves memtable writes to the flush thread, after a imm
wal segment has been flushed. This way, only committed writes enter
the memtable and its straightforward to implement committed vs
uncommitted reads by reading from the memtable only or including the
unflushed wal, respectively.