compaction: orchestrator for coordinating compaction work
This patch adds the compactor thread which orchestrates
compaction from l0 to the list of sorted runs:

The core of the compactor is the compactor_state module.
This module contains the core compactor state, and provides
a wrapping struct for reading and writing this state, called
CompactorStateHolder. CompactorStateHolder executes specific
mutations like merging updates from the writer, submitting
new compactions, and updating the core db state when a
compaction is finished. CompactorStateHolder also implements
traits that allow restricted access to the state:
CompactorStateReader allows reading compactor state only. All
   reads are by-value, meaning they return a copy or pointer to
   immutable data. So new writes will never be reflected in the
   returned data.
CompactionWriter allows submitting new compactions.

Additionally, CompactorStateHolder also allows for registering
listeners that are notified whenever the contained state changes.
For example, a listener can be notified when a new writer update
is merged, the compactor has written new db state, a new compaction
was submitted, etc.

Compaction is orchestrated by CompactionOrchestrator. It runs in
response to timer events (to periodically poll for writer updates)
or when messages come in on one of its channels. It has channels
for being notified of shutdown and for updates from compaction
workers that will actually execute compactions. (note that in this
patch we don't actually execute any compactions - the orchestrator
just rewrites the same ssts into the target sorted run).

CompactionOrchestrator also registers itself as a state listener.
When the state is updated, it writes the update to the manifest.
When new compactions are submitted it kicks off the compaction work.

Finally, this patch also adds a very simple version of the size
tiered compaction scheduler. A compaction scheduler accepts a
pointer to a CompactionStateReader and CompactionWriter and
uses these to read state and submit compactions, respectively.
Currently, it just compacts l0 when it crosses 8 SSTs.