compaction: orchestrator for coordinating compaction work
This patch adds the compactor thread which orchestrates
compaction from l0 to the list of sorted runs.

The core of the compactor is the compactor_state module.
This module contains the core compactor state in CompactorState,
and provides methods for reading state and executing specific
mutations like merging updates from the writer, submitting new
compactions, and updating the core db state when a compaction
is finished.

Compaction is orchestrated by CompactionOrchestrator. It runs in
response to timer events (to periodically poll for writer updates)
or when messages come in on one of its channels. It has channels
for being notified of shutdown and for updates from compaction
workers that will actually execute compactions. (note that in this
patch we don't actually execute any compactions - the orchestrator
just rewrites the same ssts into the target sorted run). The
orchestrator runs on a thread. It stores a handle to the tokio
runtime that it uses to execute async tasks.

Finally, this patch also adds a very simple version of the size
tiered compaction scheduler. A compaction scheduler implements a
method called `maybe_schedule_compaction` that takes a reference
to the current compactor state and returns any compactions that
should be executed. Currently, it just compacts l0 when it crosses
8 SSTs.