Remove `failpoints.rs` and add dependency on `fail-parallel` package
Per-the diescussion in #82 and #80, I've moved `failpoints.rs` out of SlateDB
and into its own crate. I would really prefer to have TiKV's `fail-rs` support
parallel execution. We'll see if they're open to a patch. In the meantime, I've
published a crate called `fail-parallel`. The source code is at:

   https://github.com/slatedb/fail-rs.

See #80 and #82 for additional context.

Closes #82