refactor: move filters into a cache in tablestore
This patch makes a "refactor" to tablestore that moves the filters
from the SsTable descriptor to a "cache" that resides inside tablestore.
The cache is not truly a cache, its just a map where we store all the
filters we've ever loaded. Filters are cached when SSTs are written,
or lazily the first time an SST is loaded. Though its not a true cache,
this patch does move us a bit closer to the desired end state, where
filters are cached in the db's in-memory/on-disk cache. The benefit
of this increment is that we no longer need to do i/o to load up the
sst descriptor (SsTableHandle) - it can be loaded from the manifest
alone. This really simplifies all the db state code that needs to load
descriptors to build up db state.

I've also moved SsTableHandle and SsTableId to the db_state module. I'd
like to rename SsTableHandle to SsTable as its really a descriptor rather
than a handle that can be closed. But I defer that to a later patch to
keep the diff smaller.

I also added a couple unit tests for db_state's refresh_db_state method
now that the dbstate is easier to load.