move sst index into its own data block (#140)
This patch moves the sst index out of the sst info footer and into its
own data block that sits just before the footer. The index can grow
quite large, and therefore we don't want to require that it be stored in
memory, or put it in the manifest. Instead, this patch changes the read
path to load the index by reading it from the sst when its needed.