support disabling the WAL (#144)
This patch adds the ability to disable the WAL. Some use cases may not
require SlateDB to maintain its own WAL and prefer not to take on the
cost of maintianing the WAL and restoring it. This patch allows clients
of the DB to disable the WAL using a db option. Because this option is a
bit risky - we wouldn't want it to be set to false by mistake - it's
only enabled when a rust feature called wal_disable is enabled.