implement compaction policy with backpressure (#147)
this patch fills in the size-tiered compaction policy and adds a crude
backpressure mechanism to throttle writes if they are arriving faster
than the db can compact:

Updates SizeTieredCompactionPolicy so that it compacts sorted runs under
compacted. The policy works by looking for sequences of runs that are
all similarly sized. Similarly sized is determined by an option to the
policy called include_size_threshold. A SR is included in a given
sequence if it's size is less than include_size_threshold multiplied by
the size of the smallest run that is already in the sequence.

The policy eliminates sequences that violate the following constraints:
- conflicts: A compaction is eliminated if it includes a source for
which there is already a compaction running.
- backpressure: A compaction is eliminated if it produces a run that is
adjacent to lots of other runs that are also of similar size to the
produced run (estimated using the sum of the sizes of the runs in the
compaction). The threshold for the number of similarly-sized runs is the
size of the max compaction (defaults to 8). This way, we won't schedule
this compaction until those larger runs are compacted.

This patch also includes a mechanism for applying backpressure to
writes. The compaction policy applies backpressure by waiting to compact
when there are too many larger runs. The writer then applies
backpressure by pausing memtable flushes and writes, based on the
following options:
- l0_max_ssts: if there are more l0 SSTs than this value, the writer
won't flush immutable memtables.
- max_unflushed_memtable: if there are more immutable memtables than
this value, the writer pauses writes until the memtables are flushed.

Currently writes block until the backpressure condition is no longer
met. In the future we could opt to instead simply slow down the writes
to some rate limit, or return an error and let the client back off.

With this patch, using the db bench tool in #148, I was able to load
~500GB into a database on an m5.xlarge (4 cores, 16GB memory) with
32-byte keys + 256-byte values at a rate of ~70K writes/second. The
number of runs in the database was always under 15 runs, so we should be
able to get pretty reasonable read performance (I'll measure that next).