fix(cache): oom when enabled object store cache (#211)
after this PR was applied, the script
(https://gist.github.com/flaneur2020/bd03c6d9032828c379a86c7ddeb8db32)
successfully finished after approximately 15 minutes with a peak RES of
about 500+ MB.

~it appears that this memory usage is not actually a leak, but rather a
reserved memory buffer from the memory allocator. this buffer tends to
grow and is not released to the OS immediately. the earlier code
allocated the buffer in part_size, which might not have been very
friendly to the allocator's buffer management.~

from @rodesai :

> I don't think it's a problem with the allocator's buffer management
per se. I think that's happening is that we were returning the Bytes
returned from slice here, which doesn't allocate a new buffer, but
instead just adds a reference to the underlying buffer that stores the
whole range and wraps that in the relevant offset+range. We then go and
cache those Bytes in the in-memory cache (after slicing it a couple more
times). So the moka cache thinks it has 4K blocks, but they are really
4K blocks that point to slices in 64MB ranges, and we blow up the memory
usage from the in-memory cache.


this PR added a range_in_part feature, allowing it to read only the
needed range in the part, rather than reading the entire part file into
memory.

this PR also adds a `fsync` on atomic writes in the fs, which may help
reduce the dirty pages floating in the memory, and reduce the risks of
file corruptions.