first pass at slatedb read-manifest CLI (#212)
This is the first step to implementing and Admin CLI, as the result of
the discussion on #151 and the discord channel. Note that this specific
PR does not spin up an entire embedded SlateDB instance and instead just
directly leverages the ability to read the manifest, this makes it
really fast!