Remove `SsTableInfoOwned` and add `SsTableInfo` (#222)
We've been passing FlatBuffer objects around for SST info data. We've
been wrapping these objects in `SsTableInfoOwned` because FlatBuffer's
code generated structs don't allow borrwing easily. This leads to some
annoying side effects:

- We're using FlatBuffer APIs all over the place
- We're using `.borrow()` everywhere

The Manifest doesn't work this way. Instead, we have a normal Rust
struct for Manifest and we have a codec to convert back and forth
between the FlatBuffer Manifest and our own. This PR makes SsTableInfo
behave like Manifest.

The drawback to our new approach is that we don't lazy decode the
FlatBuffer like we used to. This could impact performance in some cases.
In practice, I suspect it'll be negligible. Still, we should run some
tests to verify.

Fixes #213