feat(filter): implement hash sharing to reduce filter cost (#223)
Implement _hash sharing_ in [Reducing Bloom Filter CPU Overhead in
LSM-Trees on Modern Storage
Devices](https://dl.acm.org/doi/10.1145/3465998.3466002)

The idea is very simple and intuitive. In the point query of LSM, we
need to probe the filter. Since the key remains unchanged, we only need
to calculate the hash value once and pass the hash value to filter. So
we can reduce the hash overhead of the filter during each point query.
![屏幕截图 2024-09-21
164154](https://github.com/user-attachments/assets/387ce8be-02a0-4784-9e78-db317884dfdb)


In addition, we may be able to implement hash sharing during the
construction process of SSTable. For example, in levels below L0, the
same key may be arranged together with different values. In SSTable
Builder we can use a field named `last_key_hash`, and every time a new
key is added, we compare it with `last_key`. If it is the same, we can
avoid the hash cost required to build the filter during compaction.

We can also use a BTreeMap to save the existing key and its hash value
since the block size is fixed, we can estimate the space overhead it
brings(use BTreeMap instead of `last_key_hash`). But I'm not sure if
slateDB welcomes this idea, if allowed I'd be happy to experiment with
it. At least I think it's worth using a `last_key_hash` field and only
paying for a u64 field and simple comparison to avoid unnecessary hash
costs.

---------

Co-authored-by: Chris Riccomini <criccomini@apache.org>