Use ulid, no crash protection, single flush thread (#220)
Making some changes to actually reflect the design.

1.  Use chronologically sortable id for L0 names (to model ULID)
2.  Does not handle flush thread crashes (`crash_on_yield: false`)
3.  Support only a single flush thread (`max_concurrent_actions: 1`)
4.  Add assertion that uncommitted read always returns last put value
5. Add assertion that committed read eventually returns the last put
value
6.  Add an assertion MaxObjects as safety check. 
- The MaxObjects in the ObjectStore should be at most 2*number of Puts.
    -  Violating 2 and 3 above will fail MaxObjects check

There is a small bit that deviates from the actual design. The FlushWal
and FlushMemtable are modeled as separate actions, instead of FlushWal
triggering FlushMemtable as an option. One implication is there could be
two threads one flushing wal and one flushing memtable at a time.