chore(cache): Replace CachedBlock with CachedEntry.
Stop using an enum to insert data in the cache. This allows to keep the implementation hidden without exposing internals like SsTableIndexOwned.

Change the DbCache interface to receive and return the same data structure. This is more consistent and easier to implement if someone wants to implement their own cache.