Use macros instead of if's for block cache config
I noticed that we fail with a compiler error if we build without the foyer
feature. It looks like the if statements in our block cache config are still
compiled even if the feature is not enabled. I switched to use a macro instead
to prevent compiling foyer code when the foyer feature is not enabled.