Add simple benchmarks to BloomFilterBuilder
There are benchmarks for normal bloom filter operations
(building, querying, and insertion). However, the main benchmark of
interest in this commit is bloom_filter_insert_benchmark, as this
allows us to stress filter_hash(). As the flmaegraph indicates, we
spend roughly 12% of benchmark execution time in filter_hash().

The hope is to reduce this in subsequent PRs/patches