[1/N] RFC-3: introduce row_attributes into row format for tombstones
This commit adds the concept of row attributes per SST. For now, every
SST implicitly has row_attributes of [RowFlags], where the RowFlags
attribute indicates that the first byte after the key_suffix indicates
the flags associated with the row. For now the only flag is the presence
of a tombstone, in the first bit of the bitflags.

This sets us up for a future PR where we introduce the row_attributes of
timestamp and expire_ts, which will be optionally encoded with each row in
cases where TTL is enabled.

This commit also has a minor refactor of the codec so that decoding and
encoding rows happen in the same file (ease of read/testing).