[1/N] RFC-3: introduce row_attributes into row format for tombstones (#251)
This commit adds the concept of row attributes per SST. For now, every
SST implicitly has row_attributes of [RowFlags], where the RowFlags
attribute indicates that the first byte after the key_suffix represents
the flags associated with the row. For now the only flag is the presence
of a tombstone, in the first bit of the bitflags.

This sets us up for a future PR where we introduce the row_attributes of
timestamp and expire_ts, which will be optionally encoded with each row
in cases where TTL is enabled.

This commit also has a minor refactor of the codec so that decoding and
encoding rows happen in the same file (ease of read/testing).

----

## Review Guide

- Start with the new file `row_codec.rs`, which includes the new
encoding/decoding logic. Note that I've included `v0` in case we ever
change the row format (similar to this PR) and then we can encode that
into the SsTableInfo
- Check `flatbuffer_types.rs` which is where the `row_attributes` are
set in production. Note that we are currently not encoding the
`row_attributes` in the SST files themselves (as indicated in RFC-3)
since this proved to be a little tricky (that'll be in the next PR)
- `block_iterator.rs` now handles possible errors from decoding invalid
`row_attributes`
- The rest is plumbing!