Add `WriteBatch` and `write_batch` operations for atomic batch writes
Taking a crack at a `WriteBatch` API similar to RocksDB's. I opted to use a
`WriteBatch` struct rather than an iterator of key-value pairs so that we could
record deletes as well.

I believe this should "just work" since we hold a guard for the db state in each
of these operations:

- write batch
- put
- delete
- WAL flush
- memtable flush