Move `compaction_execute_bench` into `bencher` tool (#261)
#247 merged `db_bench` and `s3_bench` into one tool called `bencher`.
This PR furthers that effort by moving `compaction_execute_bench` into
the `bencher` tool as well. A new subcommand called `compaction` is
added to run the benchmarks. The `compaction`, subcommand has 3
subcommands: `load`, `run`, `clear`. These commands mirror the previous
version of the tool.

Some notes:

- Uses CLAP for args instead of the old `Option` and `parse_*` code.
- Uses `admin.rs` to load `object_store` objects (and thereby supports
`.env`).
- I noticed the previous version of the tool hard coded all destinations
to a `0` ID even though it had a `destination` parameter. I chose to
default to 0, but use the `destination` argument if it was specified.