Parameterize `cargo run --bin bencher`
While testing #264, I noticed that the bencher script writes all of its tests into the same object store location. This causes sequential runs to see the WALs an manifests of prior runs.

This PR parameterizes the bencher command and defaults each object path to include the put percentage and concurrency.