Add `WriteBatch` and `write_batch` operations for atomic batch writes (#264)
Taking a crack at a `WriteBatch` API similar to RocksDB's. I opted to
use a `WriteBatch` struct rather than an iterator of key-value pairs so
that we could record deletes as well.

I believe this should "just work" since we hold a guard for the db state
in each of these operations:

- write batch
- put
- get (read guard only)
- delete
- WAL flush
- memtable flush

Note that the `WriteBatch` has an unlimited size. This means that batch
writes can exceed `l0_sst_size_bytes` (when `WAL` is disabled). It also
means that WAL SSTs could get large if there's a large batch write.