feat(db): Allow loading DbOptions from a configuration file or environment variables (#270)
This change adds these functions to allow loading DbOptions from
configuration files or environment variables:

`from_file`: loads the options from a specific file.
`from_env`: loads the options from environment variables with a specific
prefix.
`load`: loads the options with some default file locations and
`SLATEDB_` as a default prefix. This option is accumulative. If a config
file exists, and environment variables also exist, both are combined.
Environment variables override settings in the configuration file.

I was going to document all this and provide additional configuration
examples, but I could not find a good place to do it. Let me know where
I should add that @criccomini. If this repo doesn't include the
documentation, I'll be happy to open a new PR in a different repository
for the documentation.

PS: Most of the changes are in `src/config.rs` which GitHub collapses 🤷 

Fixes #263