[4/N] RFC-3: add support for ttl expiry during compaction (#282)
This PR adds the physical component of TTL in accordance to RFC-3
(#228). It functions by encoding an `expire_ts` in the row as its
written to the SST and then filters out rows during compaction that have
already expired. Some intricacies:

- added `PutOption` to the public API to have something that is applied
per-row (compared to `WriteOption` that is only applied per-write, which
may contain multiple rows)
- added a clock to the `CompactorOptions` so that the timestamp of
compaction can be set (this determines the "now" timestamp for
expiration)

See the PR sequence for more background:
- #251 
- #258 
- #265 

Follow-Ups:

- [ ] Don't encode `expire_ts` in the SST if no row in the SST has an
`expire_ts` set
- [ ] Filter out values on retrieval if they're already expired and
compaction hasn't kicked in yet
- [x] Support default TTLs when opening a db connection