Unify write commit pipe (#290)
This PR migrates all put/delete calls to use write batches (and thus,
the commit
pipeline). For single row operations, a batch with one row is created
under the
hood. The PR also includes two other changes:

### Return results for put/delete

The `write_with_options` method returns a `Result<(), SlateDbError>`.
Now that
we use this method for puts and deletes, they need to unwrap or forward
the
results. I opted to forward them, as this lays the groundwork for #111.
Note
that, though we now return a `Result`, it only returns `Ok` right now in
the
commit pipeline. #111 must forward actual errors as they occur.

### Move backpressure into the synchronous call path

I noticed that I moved backpressure into the commit pipeline (batch
write loop)
in a separate thread in #264. This works fine for tasks that have
`WriteOptions`
where `await_durable` is true, but for `await_durable` set to false, it
causes a
problem. Because `write()` calls simply put the WriteBatch into an
`UnboundedSender`, if an `await_durable` writer outpaces the commit
pipeline, it
will eventually overrun memory. I moved the backpressure check back into
`write_with_options` in `DbInner`, so this won't happen.