refactor(cached_object_store): Move FsCacheStorage into a new module. (#295)
I was looking into #177 and I started to think that it might be better
to split it into different PRs to avoid a massive change like I did with
#231.

So this first change only moves things around to have a better structure
to implement multiple storage backends for the object cache. The change
should be backwards compatible since it doesn't really change any
functionality.