Throttle WAL writes even if await_durable is false (#303)
This commit adds backpressure even in cases where the writer is not
waiting for durable writes. Before, we were applying backpressure by
capping the number of L0 memtables in object storage and capping the
number of unflushed memtables. This would mostly work, but if writers
went too fast, they could fill very large WAL SST files before 4
memtables were flushed. We experienced this most acutely when disabling
durable writes all together, as it meant writers would not even wait for
writes to be written to WAL object storage.

Now, we simply cap how much total memory we're allowed to use for
unflushed data, whether that data is in `imm_wal` or `imm_mem_table`. If
we exceed this amount of memory, we apply backpressure until one of the
unflushed tables gets flushed (whether it's in the WAL or memtable
queue).

I wrote a kind of silly test to verify that the WALs at least stack up
when they exceed `l0_sst_size_bytes`. It doesn't actually verify
backpressure is applied, but it at least mimics the memtable
backpressure test we had.

Closes #294