feat(snapshot): add seqence number to sst rows; remove SstRowFeature (#298)
this is one of a series of PRs to add snapshot support into slatedb.

this PR makes a change on the sst row format to encode the sequence
number inside it. this PR is considered as a **BREAKING CHANGE**.

all the sequence numbers are still zero in this PR, the generation &
maintaince of sequence number is planned to be seperated in another PR.

the change could be illustrated as:

```diff
-///  |--------------------------------------------------------------------------|
-///  |       u16      |      u16       |  var        | var  |    u32    |  var  |
-///  |----------------|----------------|-------------|------|-----------|-------|
-///  | key_prefix_len | key_suffix_len |  key_suffix | meta | value_len | value |
-///  |--------------------------------------------------------------------------|
-///  |---------------------------------------------------------------------------------------------------------------|
+///  |---------------------------------------------------------------------------------------------------------------------|
+///  |       u16      |      u16       |  var        | u64     | u8        | i64       | i64       | u32       |  var      |
+///  |----------------|----------------|-------------|---------|-----------|-----------|-----------|-----------|-----------|
+///  | key_prefix_len | key_suffix_len |  key_suffix | seq     | flags     | expire_ts | create_ts | value_len | value     |
+///  |---------------------------------------------------------------------------------------------------------------------|
```

we had a flexible `meta` field to store the metadata of the rows such as
tombstone, timestamp, expiration, etc. however, it might have benefits
to put `seq` a higher position:

1. `seq` is in fact a part of key, concatenating `key` and `seq` allows
us storing multiple versions of a key in the LSM, and `key` + `seq`
makes a unique identifier of a row.
2. in the iterator with sequence number, we need to construct a prefix
scan with `key` + `seq` to find the row we want, it's a must for us to
store them adjacently.

in leveldb, there's a concept called "InternalKey" which concatenates
`UserKey` (which is the key user orignally inputed), `seq`, `flags`
(which is a tombstone marker) together. this refactor upgrades the `seq`
and `flags` upwards, ~and move the optional part to a flat buffer
encoded `SstRowExtra` struct, which contains two optional fields
`create_ts` and `expire_ts`. i dunno whether it's too heavy weight to
having flat buffer here or not.~, and makes two fields `create_ts` /
`expire_ts` optional with flags `RowFlags::HasExpireTs` and
`RowFlags::HasCreateTs`.

this PR also made a refactor about renaming `KeyValueDeletable` into
`RowEntry`, it expands the fields `key`, `value`, `seq`, `flags`,
`expire_ts` into a flat struct. all these fields are frequently
accessed, a flat struct may make these fields easier to access. also, we
can directly pass `RowEntry`s between `SsTableBuilder`, `BlockBuilder`
and even MemTable instead of `(key, value, attrs)` after this refactor,
this may make it easier to keep these value consistent across several
function calls.

---------

Co-authored-by: Chris Riccomini <criccomini@users.noreply.github.com>