fix(flush): Handle background flush errors. (#321)
When a flush operation fails in the background, an error is populated in
the COWDbState structure that handles the state of the database.
Foreground operations check this error and returns it when the user
tries to read or write from the database.

This is another improvement to handle errors described in #111.