Add release.yaml and RELEASE.md (#334)
This PR adds a release.yaml Github action to make it easier to publish
SlateDB releases. The action is triggered through the Github actions
page. To run it, SlateDB committers input a version number and click
"run workflow". The workflow will:

1. Verify that the version adheres to the semantic versioning format.
2. Check out the latest code from the `main` branch.
3. Update the version in Cargo.toml to the specified version.
4. Commit the changes and push to the `main` branch.
5. Create a Github release with the specified version and auto-generated
release notes.
6. Publish a release to crates.io.

I've also added a RELEASE.md to describe this process.