Add ID to ManifestMissing error (#376)
Fixes #366. In addition to adding ID to `ManifestMissing`, this patch
creates a separate error `LatestManifestMissing` when we are trying to
find the latest and do not have a specific ID. This patch also adds a
couple helpers for the common case when we expect either the latest
manifest or a specific manifest to exist.