Ditch Buildjet for Warpbuild (#388)
Buildjet was still running out of disk. The machines might be fast
enough to fill quite a bit of disk. It appears Buildjet only has 64GB
disk runners, so I'm switching to Warpbuild. Warpbuild also has the
advantage of offering BYOC, which means we should be able to use it to
provision AWS machines as well.