Add `--clean` option for `bencher` (#393)
`bencher` now has a `--clean` argument that will delete all objects in
the input path after the `bencher` subcommand runs. This switch works
for both the compaction benchmarks and the db benchmarks.

By default, `--clean` is disabled since it's destructive. If `--clean`
is set, `bencher` will check that the object store path is empty. If it
is, it'll create a `.clean_benchmark_data` file, which will signal that
the data should be deleted after the benchmark runs. If the path is not
empty, the benchmark will fail. After the benchmark runs, if `--clean`
is set and a `.clean_benchmark_data` exists, `bencher` will remove all
objects from the path.