tighten up error handling to prevent writer clients from blocking indefinitely (#373)
This patch tries to tighten up error handling to prevent writer clients
from blocking indefinitely. Previously we could see:
- a writer could pass the initial error check and then write to the in
memory table, and then block until the table is durable. meanwhile, the
flush task could panic or become fenced and the writer would never be
notified.
- a writer could block in the backpressure loop, which never clears
because the compactor failed.

The approach here consists of the following changes:

First, we add some utility types/fns:
- `WriteOnceRegister` which represents a value that can be written once.
Later writes are dropped. The type also supports waiting for a value to
be written to the register.
- Utility methods for spawning monitored tasks and threads. A monitored
task or thread is spawned by another task/thread which waits on the
result of the monitored task/thread. Then, the monitor task/thread calls
a cleanup fn with the result.

The error field in DbImpl is removed and replaced with an error field of
type WriteOnceRegister<SlateDBError> in DbState. We had to move this to
DbState so that we could get mutual exclusion between setting the error
and creating a new writable table.

The flushed/durable notifications in the memtables are replaced with
WriteOnceRegister.

The wal flush task, memtable flush task, and compactor all spawn using
the utitlity spawn methods, and also monitor the global error register.
They signal status to the monitor task/thread by returning their status.
The cleanup fn for the wal flush task sets the global error register,
and the register in every wal table. The cleanup fn for the memtable
flush task sets the global error register, and the register in every mem
table. The cleanup fn for the comapctor sets the global error register.