Fix flaky `test_should_recover_imm_from_wal_after_flush_error` test (#420)
I started noticing that
`test_should_recover_imm_from_wal_after_flush_error` was failing
sporadically when I was testing #418. Upon some investigation, it
appears that this test has a race condition.

The test simulates a writer that fails after a write goes to the WAL but
before it is written to L0. The test then creates a new DB client, which
loads the WAL SSTs into a memtable table and freezes it using
`replay_wal`. The problem appears to be that `maybe_freeze_memtable` in
`replay_wal`:


https://github.com/slatedb/slatedb/blob/e87bc82cab78da2ac1759e53d3382fc104b6f64b/src/db.rs#L517

Freezes the memtable, which we want, but then it tells
`mem_table_flush.rs` to flush the immutable memtable:


https://github.com/slatedb/slatedb/blob/e87bc82cab78da2ac1759e53d3382fc104b6f64b/src/db_common.rs#L18-L21

Since the second DB client does not use an `fp_registry`, the
`mem_table_flush.rs` loop might pick up the
`MemtableFlushThreadMsg::FlushImmutableMemtables` message, flush the
memtable, and pop the imm_memtable off the back in `db_state.rs`'s
`move_imm_memtable_to_l0`.

To fix this, I paused the imm_memtable flushes to prevent the
`mem_table_flush.rs` loop from popping the immutable memtable until we
can take a snapshot of the DB state to do our checks.